refactor: extract SettingsPanel presentational widgets
Move the seven self-contained widget components (KeybindRow, SettingItem,
DataActionItem, AiProviderSwitch, CloudDashboard, CanvasInputModeSwitch,
PreviewModeSwitch) into src/components/panel/settings/widgets.tsx and prune
the now-dead imports. These widgets touch no parent state, so the cut is
mechanical. SettingsPanel.tsx drops from ~2415 to ~2032 lines.
